discovery: only set the same foundation if the address matches
Without this, nice will generate candidate pairs with the same
base address but with different addresses or multiple candidate pairs with
the same foundation. This causes various problems.

They cannot be identified uniquely in farstream. As a result, some may be
dropped and others are duplicated.

Skype for business (with sipe) causes problems if the two components with
the same foundation have different addresses.