build: make prefix option accept a list of strings
The parameter for the ignored network interface prefix build option
accepts a comma-separated string now. This list will be converted to a
comma-separated list of string literals for the C code.

Disable the feature:

   ./configure ...
   ./configure --with-ignored-network-interface-prefix= ...
   meson setup -D ignored-network-interface-prefix= ...
   meson setup -D ignored-network-interface-prefix=[] ...

Ignore interfaces whose names start with "virbr":

   ./configure --with-ignored-network-interface-prefix=virbr ...
   meson setup -D ignored-network-interface-prefix=virbr ...

Ignore interfaces whose names start with "virbr" or "veth":

   ./configure --with-ignored-network-interface-prefix=virbr,veth ...
   meson setup -D ignored-network-interface-prefix=virbr,veth ...