Add unit test for early closing of agent
This test verifies that if the completion callback provided to
nice_agent_close_async() has been invoked it is safe to shut down the
nice agent thread without risk of leaking the agent GMainContext.
Leaking the GMainContext involves leaking an "eventfd" file descriptor
and since the number of open file descriptors is limited (typically
1024) the application may quickly run out of file descriptors in the
presence of a leak.

The challenge mainly has to do with the way address resolution works in glib.
The asynchronous version of name lookup involves a worker thread and a pool
of threads to handle the underlying call to getaddrinfo(), which is a
synchronous call. There is a task related to the address lookup and that task
holds a reference to the agent GMainContext, which means the task has to
finish before the thread running the mainloop can be stopped.

This test is pretty much a white-box test, i.e. we know the underlying glib
machinery is making use of getaddrinfo() and therefore we instrument
getaddrinfo() in this test (see slow-resolver.c). If the underlying machinery
changes, this test will have to be revisited.