test-gstreamer: Deliver buffers to nicesink via appsrc
The main thread should in general not be used to push data to GStreamer
elements since buffers (especially large buffer lists) can take a long
time to process. Additionally, if processing a buffer list depends on
e.g. IO callbacks that are delivered to the main thread, processing will
hang forever since messages on the mean thread can't be processed while
we push buffers.

This can actually already happen if the nicesink is configured to be
reliable, because the nice_socket_set_writable_callback() mechanism
relies on main thread processing for delivery. And an upcoming commit
will use the same mechanism for the non-reliable nicesink case.

(We disregard the fact that a NiceAgent can be configured to use a
custom GMainContext since that comes with its own set of drawbacks.)