agent: Don't send unreliably over TCP if we are reliable
If agent is reliable, we must make sure to not drop packets if we run
over TCP. More specifically, we must make sure to not send messages
non-reliably. So change an if condition to ensure we always pick
`nice_socket_send_messages_reliable()` and never the unreliable
`nice_socket_send_messages()` if `agent->reliable`. This code path is
used if the underlying transport is TCP. Note that for TCP we have a
queue of messages, and if we are not reliable we prefer to drop packets
over adding packets to the TCP queue since it does not have an upper
bound. That's why we can't use plain `nice_socket_send_messages()` if we
are reliable.

Also add a regression test that fails without the fix.

How to reproduce
----------------

To reproduce the problem, run the test without the fix. Not all sent
seqnums are received even though we run in reliable mode:

    git checkout HEAD^ -- agent/
    meson compile -C build && meson test -C build test-gstreamer -v
    # Will give error: assertion failed (seqnum_errors == 0): (1666 == 0)

If you restore the fix, the test reliably passes. That is, all sent
seqnums were also received.

    git checkout HEAD -- .
    meson compile -C build && meson test -C build test-gstreamer -v
    # Test OK!