discovery: Remove TURN allocations in a consistent way
When removing a TURN allocation due to its associated candidate being
pruned, this now works in the exact same way as if the TURN allocation
was removed when closing agent the agent. This means that it may apply
retransmits to the deallocation request (zero lifetime request). It also
means that the same cleanup is applied. The old cleanup for the prune
case was to eager to close the socket, which meant there were cases
where we tried to send to that socket when it was partly deleted.

(cherry picked from commit 7538e3b87566a301d087ac138daef44dc011399f from
https://gitlab.freedesktop.org/libnice/libnice/-/merge_requests/283)