test-turn: Reuse turn credentials, stop using pre-hashed password
Using a pre-hashed pass is harder to understand and maintain and makes
re-using code (which we will do more of in upcoming commits) more
complicated.

If the password does not begin with "0x" it will be hashed
automatically. For reference, the code in coturn for that check is found
here:
https://github.com/coturn/coturn/blob/b523616b1f70311da5cf4fbe82eb9e603ae3c313/src/apps/relay/userdb.c#L754