tests: Assign G_LOG_DOMAIN=libnice-instrument-send to instrument-send.c
So that it is easy to confirm that EWOULDBLOCK is injected and roughly
to see how often it happens. It will be especially useful in upcoming
commits where we will start injecting them more randomly to avoid
problems with "resonance frequencies" where very regular EWOULDBLOCK
injections causes some code (notably for TURN traffic) to never recover.

To look at these log prints, do like this:

    G_MESSAGES_DEBUG=libnice-instrument-send meson test -C build test-gstreamer -v