test-gstreamer: Avoid race waiting for "ready ready"
If `g_main_loop_quit ()` is called before we get to `g_main_loop_run ()`
the test is stuck forever. I don't think it can happen as the code is
now but in an upcoming commit we will wait with
`g_main_context_iteration ()` for all candidates to gather before we set
candidates, and then there is a race. So avoid that race.